requirements.txt
boto3==1.20.23
Flask==2.0.0
python-dotenv==0.19.2
supervisor==4.2.2
gunicorn==20.1.0
pynamodb==5.1.0
python-dotenv==0.19.2
Flask-Session==0.4.0